Mate `nodCount` `vector<double>` for the mapping scheme using shape functions and wts

+++ src/include/Models/TransportModel.h
@@ -130,7 +130,7 @@ int nSteps;         /// @brief Number of steps to apply the load.
 double dt;          /// @brief Time increment.
 
 T_nodFlux nodFlux;        /// @brief Nodal flux.
-vector<int> nodCount;     /// @brief Counter for integration points surrounding nodes.
+vector<double> nodCount;     /// @brief Counter for integration points surrounding nodes.
 
 // PETSc ------------------------
